refactor(slice-4): convert first connectshyft route family to thin router
