+++ apps/connectshyft-api/src/modules/connectshyft/handlers/getConnectAvailability.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
-import { success } from '../../platform/envelopes/response';
-import { CAPABILITIES } from '../../platform/rbac/capabilities';
+import { success } from '../../../platform/envelopes/response';
+import { CAPABILITIES } from '../../../platform/rbac/capabilities';
 import { evaluateConnectShyftCapability } from '../featureFlags';
 import {
   requestHasAnyCapability,
